requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Environment variables
python-dotenv==1.0.1

# PDF parsing
pymupdf==1.24.7

# Data processing
pandas==2.2.2
numpy==1.26.4

# Machine Learning
scikit-learn==1.5.1
sentence-transformers==2.7.0
torch>=2.2.0
transformers>=4.41.0

# Database
mysql-connector-python==9.0.0

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# LLM API
groq==0.9.0